+++ gdsfactory/tests/test_read_gds_with_hierarchy.py
@@ -11,9 +11,8 @@ gdspath = gf.CONFIG["gdsdir"] / "mzi2x2.gds"
 
 def test_read_gds_hash2() -> gf.Component:
     c = gf.import_gds(gdspath)
-    assert (
-        c.hash_geometry() == "021d5b67cc60a49ff09fe5f1a91e116cccb75284"
-    ), c.hash_geometry()
+    h = "779f69657f4d1ac80cb1352c3a1e8e78e41c6db1"
+    assert c.hash_geometry() == h, f"h = {c.hash_geometry()!r}"
     return c
 
 
@@ -71,10 +70,10 @@ def _write() -> None:
 
 
 if __name__ == "__main__":
-    _write()
-    test_read_gds_equivalent2()
+    # _write()
+    # test_read_gds_equivalent2()
 
-    # c = test_read_gds_hash2()
+    c = test_read_gds_hash2()
     # c.show(show_ports=True)
     # test_mix_cells_from_gds_and_from_function2()