+++ gdsfactory/tests/test_read_gds.py
@@ -8,7 +8,7 @@ import gdsfactory as gf
 def test_read_gds_hash() -> gf.Component:
     gdspath = gf.CONFIG["gdsdir"] / "straight.gds"
     c = gf.import_gds(gdspath)
-    h = "3c03925d5918cbd6458f995c7703b298920d98e4"
+    h = "c8b69d8eb2f61eeb2a6600c27d83c227d1c3ce62"
     assert c.hash_geometry() == h, f"h = {c.hash_geometry()!r}"
     return c
 
@@ -41,7 +41,6 @@ def test_mix_cells_from_gds_and_from_function() -> None:
 
     when cells loaded from GDS and have the same name as a function with
     @cell decorator
-
     """
     gdspath = gf.CONFIG["gdsdir"] / "straight.gds"
     c = gf.Component("test_mix_cells_from_gds_and_from_function")
@@ -59,11 +58,11 @@ def _write() -> None:
 
 
 if __name__ == "__main__":
-    _write()
+    # _write()
 
     # test_mix_cells_from_gds_and_from_function()
-    test_read_gds_equivalent()
-    # test_read_gds_hash()
+    # test_read_gds_equivalent()
+    test_read_gds_hash()
 
     # c1 = gf.components.straight(length=1.234)
     # gdspath = gf.CONFIG["gdsdir"] / "straight.gds"